move testing sort to testSorting.js in tests folder
